One more

+++ src/layer/internal/mod.rs
@@ -663,7 +663,7 @@ impl Layer for InternalLayer {
             let mut corrected_id = id;
             if let Some(parent) = current_layer.immediate_parent() {
                 parent_count -= current_layer.predicate_dict_len() as u64;
-                if corrected_id > parent_count as u64 {
+                if corrected_id > parent_count {
                     // subject, if it exists, is in this layer
                     corrected_id -= parent_count;
                 } else {
